dashboard changed to server component
