+++ components/Dashboard/TransactionTable.tsx
@@ -1,10 +1,10 @@
 "use client";
 
+import { Transaction, MatchedPair } from "@/types/transaction";
+import { downloadCSV } from "@/utils/downloadCsv";
+import { downloadPDF } from "@/utils/downloadPdf";
+import { Calendar, ArrowUpDown, DollarSign } from "lucide-react";
 import React, { useState } from "react";
-import { Transaction, MatchedPair } from "../types/transaction";
-import { ArrowUpDown, Calendar, DollarSign } from "lucide-react";
-import { downloadCSV } from "../utils/downloadCsv";
-import { downloadPDF } from "../utils/downloadPdf";
 
 interface TransactionTableProps {
   title: string;
